requirements.txt
aiohttp==2.0.7
appdirs==1.4.3
APScheduler==3.3.1
async-timeout==1.2.1
cffi==1.10.0
chardet==3.0.3
discord.py==0.16.8
gevent==1.2.1
greenlet==0.4.12
gui2py==0.9.6.1
lxml==3.8.0
multidict==2.1.5
numpy==1.13.1
olefile==0.44
packaging==16.8
Pillow==4.2.1
praw==4.5.1
prawcore==0.10.1
pycparser==2.17
PyNaCl==1.1.2
pyparsing==2.2.0
pytz==2017.2
requests==2.14.2
six==1.10.0
tzlocal==1.4
update-checker==0.16
websockets==3.3
xmltodict==0.11.0
yarl==0.10.2
youtube-dl==2017.5.23
